[기능 구현] 김진도 - 퀴즈 성적 처리 안정화
지석범 - 확인

+++ src/main/java/com/samsamgyeesam/studyingvally/domain/quiz/dto/QuizAttemptDTO.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class QuizAttemptDTO {
-    private Integer quizScore;
+    private Long quizScore;
     private String quizResult;
     private String quizNo;
     private Long userNo;
